Applied MacWen's "header read size 0" fix.
git-svn-id: http://forge.voodooprojects.org/svn/chameleon/trunk@136 b5af796a-00a8-11df-919b-ffff7a100b7d

+++ i386/libsaio/hfs.c
@@ -887,7 +887,8 @@ static long ReadExtent(char * extent, uint64_t extentSize,
         readOffset = ((blockNumber - countedBlocks) * gBlockSize) +
                      (offset % gBlockSize);
     
-        readSize = GetExtentSize(currentExtent, 0) * gBlockSize - readOffset;
+		// MacWen: fix overflow in multiplication by forcing 64bit multiplication
+        readSize = (long long)GetExtentSize(currentExtent, 0) * gBlockSize - readOffset;
         if (readSize > (size - sizeRead)) readSize = size - sizeRead;
 
         readOffset += (long long)GetExtentStart(currentExtent, 0) * gBlockSize;
